fix(web-app): new page tabs component cleanup
Closes #1292
