--- services/web-app/components/page-tabs-2/index.js
@@ -1,9 +1,0 @@
-/*
- * Copyright IBM Corp. 2022, 2022
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-import PageTabs2 from './page-tabs-2'
-
-export default PageTabs2
