--- services/web-app/components/page-tabs-2/page-tabs-2.module.scss
@@ -1,97 +1,0 @@
-/*
- * Copyright IBM Corp. 2022, 2022
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-@use '@carbon/react/scss/breakpoint' as breakpoint;
-@use '@carbon/react/scss/spacing' as spacing;
-@use '@carbon/react/scss/theme' as theme;
-@use '@carbon/react/scss/type' as type;
-@use '@carbon/react/scss/utilities/convert' as convert;
-@use '../../styles/variables' as variables;
-@use '@carbon/react/scss/motion' as motion;
-@use 'sass:map';
-
-.container {
-  position: sticky;
-  z-index: map.get(variables.$z-indexes, 'floating');
-  top: 3rem;
-  width: 100%;
-
-  &::before {
-    position: absolute;
-    width: 100vw;
-    height: convert.rem(48px);
-    border-bottom: 1px solid theme.$layer-active-01;
-    background: transparent;
-    content: '';
-
-    @include breakpoint.breakpoint('md') {
-      right: -3rem;
-    }
-    @include breakpoint.breakpoint('lg') {
-      // subtract width of left side nav
-      width: calc(100vw - 16rem);
-    }
-    @include breakpoint.breakpoint('max') {
-      // max grid minus the width of the screen
-      right: calc(1584px - 100vw - 0.5rem);
-    }
-  }
-}
-
-.list {
-  display: flex;
-  width: 100%;
-  flex-direction: row;
-  padding: 0;
-  list-style: none;
-  scrollbar-width: none;
-}
-
-.list-item {
-  display: flex;
-
-  &:not(:first-of-type) {
-    margin-left: 1px;
-  }
-}
-
-@media screen and (prefers-reduced-motion: reduce) {
-  .link {
-    transition: none;
-  }
-}
-
-.link {
-  @include type.type-style('body-compact-02');
-
-  display: block;
-
-  height: 3rem;
-  padding: spacing.$spacing-04 spacing.$spacing-05 spacing.$spacing-03;
-  border-bottom: spacing.$spacing-02 solid transparent;
-  color: theme.$text-secondary;
-  // needed for focus state to be visible
-  outline-offset: -2px;
-  text-decoration: none;
-  //transitioning all creates a bug on focus bc of outline transition
-  transition: background motion.$duration-moderate-01 motion.$standard-easing,
-    border-top motion.$duration-moderate-01 motion.$standard-easing;
-
-  white-space: nowrap;
-}
-
-.selected-item .link {
-  // 3px accessibility minimum
-  border-bottom: spacing.$spacing-02 solid theme.$border-interactive;
-  color: theme.$text-primary;
-  @include type.type-style('heading-02');
-}
-
-.link:hover:not(:focus) {
-  border-bottom: spacing.$spacing-02 solid theme.$layer-active-01;
-  color: theme.$text-primary;
-}
