+++ services/web-app/components/page-tabs/page-tabs.module.scss
@@ -11,6 +11,7 @@
 @use '@carbon/react/scss/type' as type;
 @use '@carbon/react/scss/utilities/convert' as convert;
 @use '../../styles/variables' as variables;
+@use '@carbon/react/scss/motion' as motion;
 @use 'sass:map';
 
 .container {
@@ -20,7 +21,6 @@
 
   &::before {
     position: absolute;
-    z-index: 0;
     width: 100vw;
     height: convert.rem(48px);
     border-bottom: 1px solid theme.$layer-active-01;
@@ -41,43 +41,60 @@
   }
 }
 
-// override carbon, fix arrow right misalignment
-:global(.cds--tabs .cds--tab--overflow-nav-button--next) {
-  right: convert.rem(-16px);
-  bottom: convert.rem(1px);
+.column {
+  z-index: map.get(variables.$z-indexes, 'floating');
 }
 
-// override Carbon min height and border bottom styles
-.tab-list {
-  min-height: convert.rem(48px);
+.list {
+  display: flex;
+  width: 100%;
+  flex-direction: row;
+  padding: 0;
+  list-style: none;
+  scrollbar-width: none;
+}
 
-  :global(.cds--tabs__nav-link) {
-    max-width: fit-content;
-    padding: 0;
-    border-bottom: spacing.$spacing-02 solid transparent;
-    color: theme.$text-secondary;
-    @include type.type-style('body-short-02', true);
+.list-item {
+  display: flex;
 
-    &:hover:not(:focus) {
-      border-bottom: spacing.$spacing-02 solid theme.$layer-active-01;
-      color: theme.$text-primary;
-    }
-  }
-
-  // stylelint not recognizing class selectors in sass module global syntax
-  :global(.cds--tabs__nav-item--selected) {
-    border-bottom: spacing.$spacing-02 solid theme.$border-interactive;
-    @include type.type-style('heading-02');
+  &:not(:first-of-type) {
+    margin-left: 1px;
   }
 }
 
-.tab-link {
+@media screen and (prefers-reduced-motion: reduce) {
+  .link {
+    transition: none;
+  }
+}
+
+.link {
+  @include type.type-style('body-compact-02');
+
   display: block;
+
+  height: 3rem;
   padding: spacing.$spacing-04 spacing.$spacing-05 spacing.$spacing-03;
+  border-bottom: spacing.$spacing-02 solid transparent;
   color: theme.$text-secondary;
+  // needed for focus state to be visible
+  outline-offset: -2px;
   text-decoration: none;
+  //transitioning all creates a bug on focus bc of outline transition
+  transition: background motion.$duration-moderate-01 motion.$standard-easing,
+    border-top motion.$duration-moderate-01 motion.$standard-easing;
+
+  white-space: nowrap;
 }
 
-.tab-list :global(.cds--tabs__nav-link[aria-selected='true']) .tab-link {
+.selected-item .link {
+  // 3px accessibility minimum
+  border-bottom: spacing.$spacing-02 solid theme.$border-interactive;
+  color: theme.$text-primary;
+  @include type.type-style('heading-02');
+}
+
+.link:hover:not(:focus) {
+  border-bottom: spacing.$spacing-02 solid theme.$layer-active-01;
   color: theme.$text-primary;
 }
