+++ services/web-app/pages/libraries/[host]/[org]/[repo]/[library]/[ref]/assets/[asset]/[tab].js
@@ -13,7 +13,7 @@ import MdxPage from '@/components/mdx-page/mdx-page'
 import PageBreadcrumb from '@/components/page-breadcrumb/page-breadcrumb'
 import PageHeader from '@/components/page-header/page-header'
 import PageNav from '@/components/page-nav'
-import PageTabs2 from '@/components/page-tabs-2'
+import PageTabs from '@/components/page-tabs'
 import withLoading from '@/components/with-loading'
 import { assetTypes } from '@/data/asset-types'
 import { assetsNavData } from '@/data/nav-data'
@@ -76,8 +76,8 @@ const AssetTabPage = ({ source, tabs, assetData }) => {
           }
           <PageBreadcrumb items={breadcrumbItems} />
         </Column>
-        <Column sm={4} md={8} lg={{ start: 5, span: 12 }}>
-          {tabs && <PageTabs2 className={styles['asset-tabs']} tabs={tabs} title="asset tabs" />}
+        <Column sm={4} md={8} lg={{ start: 5, span: 12 }} className={styles['tabs-column']}>
+          {tabs && <PageTabs className={styles['asset-tabs']} tabs={tabs} title="asset tabs" />}
         </Column>
         <Column sm={4} md={8} lg={4}>
           <PageNav items={pageNavItems} contentRef={contentRef} scrollTopDistance={180} />
